+++ src/benchmark/tests/AESEncryption.java
@@ -89,7 +89,7 @@ public class AESEncryption implements Benchmark {
      * @param score the score to set
      */
     @Override
-    public void setScore(long score) {
+    public void setScore(double score) {
         Score.addToScores(getName(), score);
     }
 }
