+++ src/benchmark/tests/Compression.java
@@ -86,7 +86,7 @@ public class Compression implements Benchmark {
      * @param score the score to set
      */
     @Override
-    public void setScore(long score) {
+    public void setScore(double score) {
         Score.addToScores(getName(), score);
     }
 }
