+++ src/benchmark/tests/MD5Hashing.java
@@ -69,7 +69,7 @@ public class MD5Hashing implements Benchmark {
      * @param score the score to set
      */
     @Override
-    public void setScore(long score) {
+    public void setScore(double score) {
         Score.addToScores(getName(), score);
     }
 }
