+++ src/benchmark/tests/QuickSort.java
@@ -79,7 +79,7 @@ public class QuickSort implements Benchmark {
      * @param score the score to set
      */
     @Override
-    public void setScore(long score) {
+    public void setScore(double score) {
         Score.addToScores(getName(), score);
     }
 }
