+++ src/benchmark/tests/SHA512Hashing.java
@@ -67,7 +67,7 @@ public class SHA512Hashing implements Benchmark {
      * @param score the score to set
      */
     @Override
-    public void setScore(long score) {
+    public void setScore(double score) {
         Score.addToScores(getName(), score);
     }
 }
